Moved common functions to url helper
