+++ examples/experimental/i3s-17-and-debug/app-debug.js
@@ -1,4 +1,4 @@
-/* global window, fetch, URL */
+/* global fetch */
 import React, {PureComponent} from 'react';
 import {render} from 'react-dom';
 import {StaticMap} from 'react-map-gl';
@@ -20,7 +20,7 @@ import {getFrustumBounds} from './frustum-utils';
 import TileLayer from './tile-layer/tile-layer';
 import AttributesTooltip from './components/attributes-tooltip';
 import {getTileDebugInfo} from './tile-debug';
-import {createTilesetUrl} from './url-utils';
+import {parseTilesetUrlFromUrl, parseTilesetUrlParams} from './url-utils';
 
 const TRANSITION_DURAITON = 4000;
 
@@ -71,26 +71,6 @@ const VIEWS = [
     }
   })
 ];
-
-function parseTilesetUrlFromUrl() {
-  const parsedUrl = new URL(window.location.href);
-  return parsedUrl.searchParams.get('url');
-}
-
-function parseTilesetUrlParams(url, options) {
-  const parsedUrl = new URL(url);
-  let token = options && options.token;
-  const tilesetUrl = createTilesetUrl(parsedUrl);
-  const index = tilesetUrl.lastIndexOf('/layers/0');
-  let metadataUrl = tilesetUrl.substring(0, index);
-
-  if (parsedUrl.search) {
-    token = parsedUrl.searchParams.get('token');
-    metadataUrl = `${metadataUrl}${parsedUrl.search}`;
-  }
-
-  return {...options, tilesetUrl, token, metadataUrl};
-}
 
 export default class App extends PureComponent {
   constructor(props) {
