+++ examples/experimental/i3s-17-and-debug/url-utils.js
@@ -1,4 +1,25 @@
-export function createTilesetUrl(parsedUrl) {
+/* global window, URL */
+export function parseTilesetUrlFromUrl() {
+  const parsedUrl = new URL(window.location.href);
+  return parsedUrl.searchParams.get('url');
+}
+
+export function parseTilesetUrlParams(url, options) {
+  const parsedUrl = new URL(url);
+  let token = options && options.token;
+  const tilesetUrl = prepareTilesetUrl(parsedUrl);
+  const index = tilesetUrl.lastIndexOf('/layers/0');
+  let metadataUrl = tilesetUrl.substring(0, index);
+
+  if (parsedUrl.search) {
+    token = parsedUrl.searchParams.get('token');
+    metadataUrl = `${metadataUrl}${parsedUrl.search}`;
+  }
+
+  return {...options, tilesetUrl, token, metadataUrl};
+}
+
+function prepareTilesetUrl(parsedUrl) {
   if (parsedUrl.pathname.includes('layers/0')) {
     return parsedUrl.href;
   }
